Move registry.api.ts into projectGitLab.api.ts
